app/requirements.txt
Django>=3.2.8,<3.3.0
psycopg2-binary>=2.9.1,<2.10.0
flake8>=3.9.2,<3.10.0
djangorestframework>=3.12.4,<3.13.0
drf-yasg>=1.20.0,<1.21.0
black>=22.3.0,<22.4.0
django-redis>=5.0.0,<5.1.0
celery>=5.2.3,<5.3.0
pandas>=1.4.2,<1.5.0
django-filter>=2.4.0,<2.5.0
Pillow>=9.1.0,<9.2.0
plotly>=5.7.0,<5.8.0
kaleido>=0.2.1,<0.3.0